# docs/swagger/auth-service.yaml
openapi: 3.0.0
paths:
...
  /api/admin/users/{id}:
    get:
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags:
        - admin
      security:
        - bearer: []
    patch:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags:
        - admin
      security:
        - bearer: []
    delete:
      operationId: UsersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - admin
      security:
        - bearer: []
...
  /api/admin/roles:
    get:
      operationId: RolesController_findAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
      tags:
        - admin
      security:
        - bearer: []
    post:
      operationId: RolesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleDto'
      responses:
        '201':
          description: ''
      tags:
        - admin
      security:
        - bearer: []
  /api/admin/roles/{id}:
    get:
      operationId: RolesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags:
        - admin
      security:
        - bearer: []
  /api/admin/modules:
    get:
      operationId: RolesController_listModules
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - admin
      security:
        - bearer: []
  /api/admin/actions:
    get:
      operationId: RolesController_listActions
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - admin
      security:
        - bearer: []
  /api/admin/roles/{id}/permissions:
    patch:
      operationId: RolesController_updatePermission
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRolePermissionDto'
      responses:
        '200':
          description: ''
      tags:
        - admin
      security:
        - bearer: []
...
info:
  title: auth-service
  description: Autenticación, usuarios, roles y privilegios
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UpdateUserDto:
      type: object
      properties:
        email:
          type: string
        status:
          type: string
          enum:
            - pending
            - active
            - suspended
...
    CreateRoleDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
      required:
        - name
    UpdateRolePermissionDto:
      type: object
      properties:
        moduleSlug:
          type: string
        actionSlug:
          type: string
        allowed:
          type: boolean
      required:
        - moduleSlug
        - actionSlug
        - allowed
...